Refactor diagnostic to have own classes
