+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/MissingMethods.php
@@ -11,6 +11,7 @@ use Microsoft\PhpParser\Token;
 use Phpactor\TextDocument\ByteOffsetRange;
 use Phpactor\WorseReflection\Core\Diagnostic;
 use Phpactor\WorseReflection\Core\DiagnosticProvider;
+use Phpactor\WorseReflection\Core\DiagnosticSeverity;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\NodeContextResolver;
@@ -57,7 +58,7 @@ class MissingMethods implements DiagnosticProvider
         try {
             $name = $class->methods()->get($methodName);
         } catch (NotFound $notFound) {
-            yield new Diagnostic(
+            yield new MissingMethodDiagnostic(
                 ByteOffsetRange::fromInts(
                     $memberName->getStartPosition(),
                     $memberName->getEndPosition()
@@ -67,7 +68,9 @@ class MissingMethods implements DiagnosticProvider
                     $methodName,
                     $containerType->__toString()
                 ),
-                Diagnostic::ERROR
+                DiagnosticSeverity::ERROR,
+                $class->name()->__toString(),
+                $methodName
             );
         }
     }
